Fix for feature "expose all notes manifest attributes in '.service_nodes'"
This feature is very usefull but cause problems with large services.
Unit's files get too big and service start could fail because of that.

This PR aims to add a new yaml property 'exposeNodesInUnitEnv' as boolean
to make '.service_nodes' optional.

Set to 'true', service nodes will be exposed.
Set to 'false' (or not set), they won't